requirements.txt
mem0
groq
python-dotenv